test: move AutoTagger and PDF export tests to ArticleRightPane
PageViewPage tests for AutoTagger and PDF export removed as those
features were moved to ArticleRightPane. New tests added to
ArticleRightPane.test.tsx covering the same functionality.
